First phase of the implementation of the View Servlet action, without JSP compilation. View Servlet now displays the generated servlet that was created by the server. This means that the View Servlet action is not enabled until the page has run on the server.
The intention is that the location of the servlet for a particular page is acquired from the server plugin, if the plugin supports this. However, since now there are no server integration APIs to do this (they will be introduced when the trunk uses the new server integration APIs), the location of the generated servlet is hardcoded for Tomcat 4.0 for now.