IN=19050.
Adding APIs (on both the plugin and devmodule sides) to find the servlet
generated from a JSP. Also providing the implementation on the side of the
Tomcat plugin, and on the side of webapps-core.

The plugin-side API consists of class
org.netbeans.modules.j2ee.deployment.plugins.spi.FindJSPServlet

Also adding some initial APIs that will in the future allow JSP source level
debugging for plugins that do not support JSR 45.