Add a module cluster for the Ruby modules (ruby1). Until now, these modules have gone into extra/.
This will also add ruby1 to the set of clusters listed in etc/netbeans.clusters. This is
necessary such that when users install the Ruby support, these modules are enabled. This means
of course that ruby1 will be included for all configurations of the IDE, but the launcher will
ignore it when it is not present, and hopefully the component installer will customize the
cluster list anyway.  (Trung and Pavel have OK'd this approach).