#105212 API review request - WebLogicalViewProvider extension to support visualweb project error
For this feature, I'm adding an inner class BrokenDatasourceAction to
Web Project's WebLogicalViewProvider to support adding a Project action and badging
 for visualweb projects when database connections used by the
project are not available.

Also support added to J2EE Project Utilities to add Database connections and listen to changes to Database Connections
then remove the badged, marked project

The feature requested is similar feature as the BrokenServerSupport in
org.netbeans.modules.j2ee.common.ui

For this release, only visualweb projects are supported.  In future releases other types of projects that use Data
Sources can be supported by making a straightforward change to WebLogicalViewProvider

This enhancement is needed for the following reasoning:
If the database connections are not available then the visualweb project cannot
be developed at design-time or executed.  At design-time a Component Error
page will open instead of the expected page that has database-bound components.

Also based on feedback from UI review for the spec, see the URL field
 for a link to the spec.