Bug Fix# 107717 IllegalStateException after refactor; ide hang
IllegalStateException was happening only for page fragments. It was because during sync(),
the MarkupUnit.ensureAttributeValue() is called to ensure the value of id attribute
of the subview element is the name (sans extension) of the File. This happens
in the sync()->bind() sequence that results due to the fileRenamed() event
that is fired by Rename refactroing. The refactoring should take care of renaming the
sub view id for page fragments.

Another issue that has been fixed is that the SourceUnit was unable to remove the
PropertyChangeListener on the EditorCookie in the destroy() method after a move refactoring.
This meant that it kept on reacting to the document opened/closed property change events long
after it was logically destroyed. This was happening because it was using the old FileObject
to lookup a DataObject to in turn get the EditorCookie to remove the listener. However a null
DataObject was returned for the old FileObject. Fixed this by caching the EditorCookie in a
member field and using it to remove PropertyChangeListener. This probably started happening
after we switched to OperationListener (DataObject level) from the FileChangeListener based
management of FacesModels. This may fix some other wierd NPEs on editor close.

Added some code to protect against NPEs.

Removed code to refactor the managed-bean-class tag as the web/jsf refactoring takes care of it now.

Some preparatory code to handle web/**folder level move and rename refactoring.