Bug Fix # 107080 Rename refactoring of folders under web folder does not work for VW JSP Pages
While renaming folder containing JSP page the FacesModel for the original
JSP files lands up in a situation whereby the markup file points to the
new location (thru MarkupUnit) vs. java file points to the old location
(thru JavaUnit). While removing the original FacesModel it removes the
units. The units remove the DataObject which in turn remove the FileObjects.
Thus the markup file in the new location was getting deleted.

Added a check - before deleteing the Dataobjects - to make sure that the
files of both units in fact correspond to each other using the standard
relative path and name matching mechanism.